+++ src/dao/AdvertismentDAO.java
@@ -2,11 +2,11 @@ package dao;
 import domain.Ad;
 public interface AdvertismentDAO {
 
-	Ad getById(Long id);
+	Ad getById(int id);
 	
 	void save(Ad ad);
 	
 	void update(Ad ad);
 	
-	void deleteById(Long id);
+	void deleteById(int id);
 }
